// src/types/developer.tsx
import { JsonLangsProp } from "./common";

export const COMPOUND_RESOURCE_NAME = "developers";

export type DeveloperType = {
  created_at: string;
  description: JsonLangsProp;
  ext_data: any;
  id: string;
  media: any;
  name: JsonLangsProp;
  slug_ar: string;
  slug_en: string;
  updated_at: string;
};